requirements.txt
# ─────────────────────────────────────────────
# EventOracle – Project Dependencies
# ─────────────────────────────────────────────

# Database
pymongo>=4.6.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Market data (market_pipeline.py)
yfinance>=0.2.30

# Web scraping / HTTP (events, flows, news pipelines)
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# NLP / Sentiment (news_pipeline.py)
nltk>=3.8.0
vaderSentiment>=3.3.2

# Environment variables
python-dotenv>=1.0.0

# Timezone handling
pytz>=2024.1

# Generative AI Agent
google-genai>=0.3.0
